Added exception handlers + logs on usecases

+++ src/main/java/com/example/e_m_test/api/app/impl/client/DeleteClientEmailUseCase.java
@@ -1,9 +1,8 @@
 package com.example.e_m_test.api.app.impl.client;
 
-import com.example.e_m_test.api.app.api.client.ClientRepository;
-import com.example.e_m_test.api.app.api.client.DeleteClientEmailInBound;
-import com.example.e_m_test.api.app.api.client.EmailNotFoundException;
-import com.example.e_m_test.api.app.api.client.EmailRepository;
+import com.example.e_m_test.api.app.api.ObjectAlreadyExistException;
+import com.example.e_m_test.api.app.api.ObjectNotFoundException;
+import com.example.e_m_test.api.app.api.client.*;
 import com.example.e_m_test.api.domain.client.Client;
 import com.example.e_m_test.api.domain.client.Email;
 import lombok.RequiredArgsConstructor;
@@ -22,18 +21,25 @@ public class DeleteClientEmailUseCase implements DeleteClientEmailInBound {
     @Override
     public Client delete(Long id, Email email) {
         log.info("Client wants to delete his email: {}", email.getAddress());
-        Client client = clientRepository.getById(id);
+        Client client = clientRepository.getById(id)
+                .orElseThrow(() -> new ClientNotFoundException(id));
+
         int valueOfClientEmails = client.getEmails().size();
+
+        if (!emailRepository.existsByAddressAndClientId(email.getAddress(), id)) {
+            throw new ObjectAlreadyExistException("Email " + email.getAddress() + " does not matches to client");
+        }
 
         if (valueOfClientEmails > 1) {
             Email emailToDelete = client.getEmails().stream()
                     .filter(e -> e.getAddress().equals(email.getAddress()))
                     .findFirst()
-                    .orElseThrow(() -> new EmailNotFoundException("Email not found: " + email.getAddress()));
+                    .orElseThrow(() -> new ObjectNotFoundException("Email not found: " + email.getAddress()));
             client.getEmails().remove(emailToDelete);
             emailRepository.deleteByAddress(email.getAddress());
             return clientRepository.save(client);
         }
-        return client;
+        log.warn("Client could not to delete his last email. Rollback...");
+        throw new ObjectAlreadyExistException("You can not to delete your first email");
     }
 }
